anti-slop: replace the batch's numeric casts with conversions
Every `as` between numeric types the batch added goes through a named
conversion, or is gone:

- agent_panel: a card's `action_index` is a `u32` on the wire, so the
preview takes a `u32` and the offer is numbered with a `u32` counter
instead of `enumerate`; a request's index becomes a `usize` through
`try_from`, with a miss answered as `UnknownAction` like any other index
the turn does not have.
- context/windows: the image buffer's length becomes the `u32` the call
wants through `try_from`, sized before the handle is open so a miss cannot
leak it.
- ledger: turns per minute reads both counts through `u32::try_from`
pinned at `u32::MAX`, which `f64::from` then holds exactly (saturating on
purpose: no recorded meeting comes near either bound).
- ledger_evals: the fixture speaker id comes from `u128::try_from`, pinned
at `u128::MAX` the same way.
- session: the WAV fixture counts samples in a `u16`, which `f32::from`
holds exactly, so the tone's phase is the same bits.
- workflow_core_tests: segments are numbered by a `u64` counter zipped
with the lines.
- card: the four counts that came from `len()` and `count()` are `usize`
fields now; nothing needed them as `u64`, and the 72-hour horizon is added
through `chrono::Duration` rather than a widened millisecond constant.
- tools: the activity window's `take` goes through `usize::try_from`,
pinned at `usize::MAX` (saturating: "every row" when the count does not
fit, which it always does).

`import_formats::seconds_to_ms` keeps its `f64 as u64`: it truncates a
guarded finite, non-negative number of seconds to milliseconds and no
lossless form exists.

+++ src-tauri/src/meeting/session.rs
@@ -5581,8 +5581,9 @@ pub(crate) mod tests {
     }
 
     /// A complete 16 kHz mono WAV. The decoder verifies a WAV's declared frame
-    /// count against what it decodes, so the fixture has to be whole.
-    fn write_mono_wav(path: &Path, samples: usize) {
+    /// count against what it decodes, so the fixture has to be whole. The
+    /// count is a `u16`, which an `f32` holds exactly, so the tone's phase is.
+    fn write_mono_wav(path: &Path, samples: u16) {
         let specification = hound::WavSpec {
             channels: 1,
             sample_rate: 16_000,
@@ -5591,7 +5592,7 @@ pub(crate) mod tests {
         };
         let mut writer = hound::WavWriter::create(path, specification).expect("create wav");
         for index in 0..samples {
-            let phase = index as f32 / 16_000.0 * std::f32::consts::TAU * 220.0;
+            let phase = f32::from(index) / 16_000.0 * std::f32::consts::TAU * 220.0;
             writer
                 .write_sample(phase.sin() * 0.4)
                 .expect("write sample");
